Work on train_visual_similarity_own.py
work on engine.py, visual_sim_engine.py, ...
in engine.py split set up into test_only and train
engine.py save_model creates dict not set
engine.py call writer.flush() after adding infos

Added gradients for torch tensors
IDs are torch.long (int64) instead of torch.int (int32) to be able to
    compute losses
Due to size of PT21 dataset, have to use custom cosine similarity
Use torch cdist instead of custom euclid distance
more helpers and structure for posetrack21.py
Differentiated between BaseModule.print() and .can_print()
Add image crops path to DataSample during PT21 Dataset loading.
Add crop_path property to DataSample
Added PBEG validation
Visual models in torchreid.py are nn.Modules and use nof_classes
in loader.py removed redundant paths and used more generic versions
in utils.py extract_crops_from_images() convert the image to byte
    before saving it
Added more type validations

Docs added scripts.rst to explain scripts, modified installation.rst

Some cleanup

+++ tests/models/test__data_sample.py
@@ -119,7 +119,7 @@ class TestDataSample(unittest.TestCase):
 
     @test_multiple_devices
     def test_init_with_device(self, device: Device):
-        out_id = torch.tensor(10).int().to(device=device)
+        out_id = torch.tensor(10).long().to(device=device)
         out_image = DUMMY_IMG.to(device=device)
         out_imgcrop = DUMMY_IMG.to(device=device)
         out_kp = DUMMY_KEY_POINTS.to(device=device)
